requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy
pandas
opencv-python
matplotlib
scikit-learn
albumentations
tqdm
seaborn
Pillow
streamlit
labelme